+++ weave/__init__.py
@@ -2,7 +2,7 @@
 
 from weave import version
 
-from weave.trace.api import *
+from weave.trace import *
 from weave.flow import *
 from weave.trace.util import ThreadPoolExecutor, Thread
